generate test data in build scripts
this is ready to auto-generate on the CI server if cpp/java arrow libs are available
