+++ js/gulpfile.js
@@ -19,11 +19,11 @@ const del = require('del');
 const gulp = require('gulp');
 const path = require('path');
 const { Observable } = require('rxjs');
-const testsTask = require('./gulp/test-task');
 const buildTask = require('./gulp/build-task');
 const cleanTask = require('./gulp/clean-task');
 const packageTask = require('./gulp/package-task');
 const { targets, modules } = require('./gulp/argv');
+const { testTask, createTestData, cleanTestData } = require('./gulp/test-task');
 const {
     targetDir,
     taskName, combinations,
@@ -35,8 +35,8 @@ const {
 for (const [target, format] of combinations([`all`], [`all`])) {
     const task = taskName(target, format);
     gulp.task(`clean:${task}`, cleanTask(target, format));
-    gulp.task( `test:${task}`, testsTask(target, format));
-    gulp.task(`debug:${task}`, testsTask(target, format, true));
+    gulp.task( `test:${task}`,  testTask(target, format));
+    gulp.task(`debug:${task}`,  testTask(target, format, true));
     gulp.task(`build:${task}`, gulp.series(`clean:${task}`,
                                             buildTask(target, format),
                                             packageTask(target, format)));
@@ -86,9 +86,9 @@ const buildConcurrent = (tasks) => () =>
             .merge(...knownTargets.map((target) =>
                 del(`${targetDir(target, `cls`)}/**`)))));
   
-gulp.task( `test`, gulp.series(getTasks(`test`)));
-gulp.task(`debug`, gulp.series(getTasks(`debug`)));
-gulp.task(`clean`, gulp.parallel(getTasks(`clean`)));
+gulp.task( `test`, gulp.series(/*createTestData,*/ getTasks(`test`)/*, cleanTestData*/));
+gulp.task(`debug`, gulp.series(/*createTestData,*/ getTasks(`debug`)/*, cleanTestData*/));
+gulp.task(`clean`, gulp.parallel(/*cleanTestData,*/ getTasks(`clean`)));
 gulp.task(`build`, buildConcurrent(getTasks(`build`)));
 gulp.task(`default`,  gulp.series(`build`, `test`));
